telegram-bot/requirements.txt
python-telegram-bot==20.7
aiohttp==3.9.1
redis==5.0.1
PyMySQL==1.1.0
PyJWT==2.8.0
python-dotenv==1.0.0
asyncio-mqtt==0.16.1